:hammer: minor code refactoring
